+++ opentelemetry-api/CHANGELOG.md
@@ -2,12 +2,20 @@
 
 ## Unreleased
 
+## 0.9b0
+
+Released 2020-06-10
+
 - Move stateful from Meter to MeterProvider
   ([#751](https://github.com/open-telemetry/opentelemetry-python/pull/751))
 - Rename Measure to ValueRecorder in metrics
   ([#761](https://github.com/open-telemetry/opentelemetry-python/pull/761))
+- Adding trace.get_current_span, Removing Tracer.get_current_span
+  ([#552](https://github.com/open-telemetry/opentelemetry-python/pull/552))
 - Rename Observer to ValueObserver
   ([#764](https://github.com/open-telemetry/opentelemetry-python/pull/764))
+- Add SumObserver and UpDownSumObserver in metrics
+  ([#789](https://github.com/open-telemetry/opentelemetry-python/pull/789))
 
 ## 0.8b0